Cast to long to workaround a bug in animal-sniffer.

+++ src/main/java/org/apache/commons/collections4/bloomfilter/hasher/DynamicHasher.java
@@ -116,7 +116,8 @@ public class DynamicHasher implements Hasher {
                     buffer++;
                 }
                 return (int) Math.floorMod(function.apply(buffers.get(buffer), funcCount++),
-                    shape.getNumberOfBits());
+                    // Cast to long to workaround a bug in animal-sniffer. 
+                    (long) shape.getNumberOfBits());
             }
             throw new NoSuchElementException();
         }
